+++ app/controllers/pg_hero/home_controller.rb
@@ -220,10 +220,6 @@ module PgHero
         "1 week" => {duration: 1.week, period: 30.minutes},
         "2 weeks" => {duration: 2.weeks, period: 1.hours}
       }
-      if @database.system_stats_provider == :azure
-        # doesn't support 10, just 5 and 15
-        @periods["1 day"][:period] = 15.minutes
-      end
 
       @duration = (params[:duration] || 1.hour).to_i
       @period = (params[:period] || 60.seconds).to_i
@@ -250,17 +246,6 @@ module PgHero
     def load_stats
       stats =
         case @database.system_stats_provider
-        when :azure
-          if @database.send(:azure_flexible_server?)
-            [
-              {name: "Read IOPS", data: @database.read_iops_stats(**system_params).map { |k, v| [k, v ? v.round : v] }, library: chart_library_options},
-              {name: "Write IOPS", data: @database.write_iops_stats(**system_params).map { |k, v| [k, v ? v.round : v] }, library: chart_library_options}
-            ]
-          else
-            [
-              {name: "IO Consumption", data: @database.azure_stats("io_consumption_percent", **system_params), library: chart_library_options}
-            ]
-          end
         when :gcp
           [
             {name: "Read Ops", data: @database.read_iops_stats(**system_params).map { |k, v| [k, v ? v.round : v] }, library: chart_library_options},